\name{similarityRichards-package}
\alias{similarityRichards-package}
\alias{similarityRichards}
\docType{package}
\title{Similarity of Richards curves}
\description{
 Values are computed for similarity of Richards curves, and plotted.
 These values are computed by 'backfitted' values of the independent variable.
}
\details{
\tabular{ll}{
Package: \tab similarityRichards     \cr
Type:    \tab Package                \cr
Version: \tab 0.4.5                  \cr
Date:    \tab 2008-09-01             \cr
License: \tab GPL Version 2 or later \cr
}
}
\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <september2008richards@badsberg.eu>
}
% 
% Emmens CW.:
% The dose response relationship for certain principles of the pituitary gland,
% and of the serum and urine of pregnancy.
% Journal of Endocr. 1940; 2:194-225. 
% 
% Reeve R.:
% Two statistical methods for estimating relative potency of bioassays.
% Biopharm Int. 2000 July; 13(7):54-60. 
% 
% Klein J, Capen R, Mancinelli R, Robinett R,
% Pietrobon FPJ, Quinn J, Schofield T.:
% Validation of assays for use with combination vaccines,
% Biologicals 1999; 27:35-41. 
% 
\references{
Richards, F.J.: A flexible growth function for empirical use.
J. Exp. Bot. 1959; 10:290-300.

Finney DJ.: Statistical methods in biological assay,
3rd ed. London: Griffin; 1978. 

Finney DJ.: Response curves for radioimmunoassay.
Clinical Chemistry 1983; 29:1562-1566. 

Schofield T.: Assay Validation.
In: Encyclopedia of Biopharmaceutical Statistics.
New York: Marcel Dekker 2000; p. 21-30. 

\code{http://en.wikipedia.org/wiki/ELISA}
}
\keyword{ package }
